chore(cli): read pg image from manifest
