+++ apps/cli/src/legacy/shared/legacy-db-image.ts
@@ -1,4 +1,5 @@
 import { Effect, type FileSystem, type Path } from "effect";
+import { dockerfileServiceImage } from "../../shared/services/dockerfile-images.ts";
 
 /**
  * Resolves the local Postgres Docker image the way Go's `config.Load` does
@@ -6,13 +7,11 @@ import { Effect, type FileSystem, type Path } from "effect";
  * pg_dump / shadow-DB container (`db dump`, declarative). Promote/extend this if
  * the full service-image resolution is ever needed.
  *
- * The image tags are baked into the Go binary via the embedded Dockerfile
- * (`pkg/config/templates/Dockerfile`, parsed into `config.Images`), so they are
- * mirrored here as constants rather than read from any file.
+ * The default PG image is read from the same embedded Dockerfile manifest Go parses
+ * into `config.Images`, so the TS port tracks Dependabot bumps in that source.
  */
 
-// `FROM supabase/postgres:17.6.1.136 AS pg` (the embedded Dockerfile `pg` stage).
-const LEGACY_PG_IMAGE = "supabase/postgres:17.6.1.136";
+const LEGACY_PG_IMAGE = dockerfileServiceImage("pg");
 // `pkg/config/constants.go:12-14`.
 const LEGACY_PG14 = "supabase/postgres:14.1.0.89";
 const LEGACY_PG15 = "supabase/postgres:15.8.1.085";
